feat: move saveAppData in fixture instead of index test file
