+++ src/services/item/plugins/app/appData/test/fixtures.ts
@@ -2,6 +2,12 @@ import { v4 } from 'uuid';
 
 import { FastifyLoggerInstance } from 'fastify';
 
+import { AppDataVisibility } from '@graasp/sdk';
+
+import { Member } from '../../../../../member/entities/member';
+import { Item } from '../../../../entities/Item';
+import { AppDataRepository } from '../repository';
+
 export const GRAASP_PUBLISHER_ID = 'publisher-id';
 
 export const MOCK_JWT_SECRET = '1234567890123456789012345678901234567890';
@@ -62,3 +68,47 @@ export const MOCK_APPS = [
 export const MOCK_LOGGER = {
   error: jest.fn(),
 } as unknown as FastifyLoggerInstance;
+
+export const saveAppData = async ({
+  item,
+  creator,
+  member,
+  visibility,
+}: {
+  item: Item;
+  creator: Member;
+  member?: Member;
+  visibility?: AppDataVisibility;
+}) => {
+  const defaultData = { type: 'some-type', data: { some: 'data' } };
+  const s1 = await AppDataRepository.save({
+    item,
+    creator,
+    member: member ?? creator,
+    ...defaultData,
+    visibility: visibility ?? AppDataVisibility.Item,
+  });
+  const s2 = await AppDataRepository.save({
+    item,
+    creator,
+    member: member ?? creator,
+    ...defaultData,
+    visibility: visibility ?? AppDataVisibility.Item,
+  });
+  const s3 = await AppDataRepository.save({
+    item,
+    creator,
+    member: member ?? creator,
+    ...defaultData,
+    visibility: visibility ?? AppDataVisibility.Member,
+  });
+  const s4 = await AppDataRepository.save({
+    item,
+    creator,
+    member: member ?? creator,
+    ...defaultData,
+    visibility: visibility ?? AppDataVisibility.Member,
+    type: 'other-type',
+  });
+  return [s1, s2, s3, s4];
+};
